Steps 9+10: Governance package split, execution independence, scale preparation
## Step 9: Governance Module Evolution

Split governance_service.py (1323 lines) into a proper governance/ package:
- governance/models.py — all dataclasses + enums (GovernanceContext,
  ResolutionTrace, GovernanceSnapshot, GovernanceDiff, etc.)
- governance/events.py — EVT_* constants + _emit_governance_event
- governance/cache.py — cache state, GovernanceCacheBackend Protocol (ISSUE-014)
- governance/dependency.py — dependency propagation via topological order
- governance/resolver.py — scope chain + visibility resolution
- governance/cleanup.py — cleanup policy resolution
- governance/execution.py — execution resolution, INDEPENDENT of visibility (ISSUE-008)
  - check_visibility=False skips the visibility gate for internal/AI-triggered calls
  - capability_execution_overrides table takes precedence always
- governance/snapshot.py — GovernanceSnapshot, GovernanceDiff, diff functions
- governance/health.py — GovernanceHealthReport, run_governance_healthcheck
- governance/writes.py — governance writes + version management
- governance/policy_engine.py — PolicyRule dataclass + evaluate_rules stub (ISSUE-033)
- governance/templates.py — GovernanceTemplate export/apply/save/load (ISSUE-034)

services/governance_service.py reduced to thin re-export shim (~50 lines).
All existing callers work unchanged via backward-compat re-exports.

Thread scope added to SCOPE_PARENT: "thread" → "channel" (ISSUE-016).
Trusted tier implemented: TRUSTED_TIER_ROLE_ID guild setting promotes "user"
tier members to "trusted" when they have the configured role (ISSUE-015).

migrations/010_capability_overrides.sql — capability_execution_overrides table.

## Step 10: Scale Preparation

- governance/cache.py: GovernanceCacheBackend Protocol for Redis swap at
  sharding time (InProcessCache is the current concrete implementation).
- core/runtime/plugin_contract.py: PluginContract Protocol + validate_plugin()
  for future plugin isolation and marketplace safety.
- governance/templates.py: GovernanceTemplate — export/apply/save/load for
  multi-server governance configuration reuse (ISSUE-034).
- migrations/011_governance_templates.sql — governance_templates +
  governance_template_applications tables.
- Rate limiting added to compute-heavy commands: economymenu (3/10s),
  daily (2/5s), work (2/5s), leaderboard (2/10s), help (3/10s),
  diagnostics (2/15s) — prevents DB pool exhaustion under abuse (ISSUE-020).
- db.py: get_all_cleanup_for_guild() added for template export.
- settings_keys.py: TRUSTED_TIER_ROLE_ID added.
