+++ disbot/governance/policy_engine.py
@@ -1,0 +1,57 @@
+"""Policy engine scaffold — ISSUE-033.
+
+Layer: top of governance stack (no other governance submodule imports this).
+Stub only — no rules authored yet. Returns None always.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+
+@dataclass
+class PolicyRule:
+    """A single governance policy rule.
+
+    Attributes
+    ----------
+    rule_id:
+        Unique identifier for this rule.
+    subsystem:
+        The subsystem this rule targets.
+    condition_type:
+        One of "member_age", "role_membership", "channel_pattern", "time_range".
+    condition_value:
+        The value to match against the condition.
+    effect:
+        Either "ENABLE" or "DISABLE".
+    priority:
+        Higher priority rules win when multiple match. Default 0.
+    """
+
+    rule_id: str
+    subsystem: str
+    condition_type: (
+        str  # "member_age", "role_membership", "channel_pattern", "time_range"
+    )
+    condition_value: str
+    effect: str  # "ENABLE", "DISABLE"
+    priority: int = 0
+
+
+@dataclass
+class PolicyRuleResult:
+    """Result from evaluate_rules()."""
+
+    rule: PolicyRule
+    matched: bool
+    effect: str  # "ENABLE" or "DISABLE"
+
+
+async def evaluate_rules(ctx: object, subsystem: str) -> "PolicyRuleResult | None":
+    """Evaluate all applicable policy rules for a context and subsystem.
+
+    Returns the winning PolicyRuleResult or None if no rules apply.
+    This is a stub — no rules are authored yet.
+    """
+    return None  # stub — no rules authored yet
