+++ disbot/utils/settings_keys.py
@@ -27,3 +27,5 @@ ACTIVE_TOURNAMENT = "active_tournament"
 # Governance (owner: governance_service)
 # Stores the REGISTRY_VERSION int at which per-guild governance was last upgraded.
 GOVERNANCE_VERSION = "governance_version"
+# Role ID whose holders are treated as "trusted" tier (ISSUE-015).
+TRUSTED_TIER_ROLE_ID = "trusted_tier_role_id"
